Add files via upload
This program is for making a Bluetooth controlled robot. It uses a HC-05 Bluetooth module.  We implemented it on an Arduino Uno. The software used for controlling the robot was downloaded from the Play Store whose link is https://play.google.com/store/apps/details?id=braulio.calle.bluetoothRCcontroller&hl=en_IN.